% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{acstabs}
\alias{acstabs}
\title{ACS tables metadata}
\format{
Tibble with 1 row per ACS table, with columns table, year, title, universe,
and n (where n is the number of variables in the table).
}
\source{
Constructed from Census documentation \url{https://www.census.gov/programs-surveys/acs/technical-documentation/table-shells.html}.
}
\usage{
acstabs
}
\description{
ACS tables metadata
}
\examples{
acstabs
}
\keyword{datasets}
