% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_tidy.R
\name{acstabs_wide}
\alias{acstabs_wide}
\title{Get number of ACS variables for all tables, selected years}
\usage{
acstabs_wide(years)
}
\arguments{
\item{years}{}
}
\value{
Tibble with tables as rows, years as columns, # of variables as cells.
}
\description{
Get number of ACS variables for all tables, selected years
}
\examples{
acstabs_wide(c(2011, 2016, 2021))
}
