% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{acsvars}
\alias{acsvars}
\title{ACS variables metadata}
\format{
Tibble with 1 row per ACS variable, with columns table, line, variable, label, release, year, title, universe,
uvariable, where uvariable is a uniform variable name (experimental) that puts an underscore in names from early years,
which do not have underscores.
}
\source{
Constructed from Census documentation \url{https://www.census.gov/programs-surveys/acs/technical-documentation/table-shells.html}.
}
\usage{
acsvars
}
\description{
ACS variables metadata
}
\examples{
acsvars
}
\keyword{datasets}
