% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_var_mismatch}
\alias{check_var_mismatch}
\title{Check for label mismatches across years}
\usage{
check_var_mismatch(table, years)
}
\arguments{
\item{table}{table name as character string}

\item{years}{vector of years to examine}
}
\value{
tibble with line, year, variable and label where mismatches occur
}
\description{
Examine multiple years for an ACS table and find instances where the label for
a line in one year does not match the label for the same line in another year.
}
\examples{
check_var_mismatch("B27003", c(2016, 2021))
check_var_mismatch("B27003", c(2011, 2016, 2021))
}
