% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_tidy.R
\name{find_acstab}
\alias{find_acstab}
\title{Find ACS tables containing a given term}
\usage{
find_acstab(searchterm, tabs = bacs::acstabs)
}
\arguments{
\item{searchterm}{String to search for in the \code{title} field of \code{tabs}.}

\item{tabs}{Tibble that contains field \code{title}.}
}
\value{
Tibble with the resulting table information.
}
\description{
Find ACS tables containing a given term
}
\examples{
find_acstab("age")
find_acstab("age", acstabs_wide(c(2011, 2016, 2021)))
}
