% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{geometry}
\alias{geometry}
\title{Geometry data for areas of interest}
\format{
Tibble with 1 row per area, with affgeoid, geoid, fullname, and geomery.
}
\source{
Constructed using tidycensus
}
\usage{
geometry
}
\description{
Geometry data for areas of interest
}
\examples{
geometry
}
\keyword{datasets}
