% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_tidy.R
\name{get_acstab}
\alias{get_acstab}
\title{Read a single ACS table and return the full table for geographies of interest}
\usage{
get_acstab(tabname, year = 2021)
}
\arguments{
\item{tabname}{Census Bureau name for the table (e.g., B01001) as a string.}
}
\value{
Tibble with the full data for the ACS table.
}
\description{
Read a single ACS table and return the full table for geographies of interest
}
\examples{
df <- get_acstab("B01003")
}
