% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_data_newformat.R
\name{read_tabny}
\alias{read_tabny}
\title{Read a previously saved enhanced ACS table from disk}
\usage{
read_tabny(tabname, xdacs = bacs:::default_xdacs)
}
\arguments{
\item{tabname}{Census Bureau name for the table (e.g., B01001) as a string.}

\item{xdacs}{Local folder where table is saved (or default).}
}
\value{
Tibble with the enhanced filtered ACS table.
}
\description{
Read an ACS table that has been enhanced with additional geographic
identification information.
}
\examples{
read_tabny("B01001")
}
