% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_data_newformat.R
\name{save_tabny}
\alias{save_tabny}
\title{Read, enhance, and save an ACS table}
\usage{
save_tabny(tabname, xdacs = bacs:::default_xdacs, overwrite = FALSE)
}
\arguments{
\item{tabname}{Census Bureau name for the table (e.g., B01001) as a string.}

\item{xdacs}{Local folder where table is saved (or default).}
}
\description{
Read, enhance, and save an ACS table
}
\examples{
save_tabny("B01003", overwrite=TRUE)
}
