% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{xwalk}
\alias{xwalk}
\title{Geographic identification variables for areas of interest}
\format{
Tibble with 1 row per area, with affgeoid, geoid, fullname, and many other identification variables.
}
\source{
Constructed using tidycensus
}
\usage{
xwalk
}
\description{
Geographic identification variables for areas of interest
}
\examples{
xwalk
}
\keyword{datasets}
