% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_tidy.R
\name{enhance}
\alias{enhance}
\title{Enhance a single ACS table and return the full table for geographies of interest}
\usage{
enhance(tabdf, year = 2021)
}
\arguments{
\item{tabdf}{Census Bureau name for the table (e.g., B01001) as a string.}

\item{year}{4-digit numeric value, must be a 5-year ACS endyear, default 2021}
}
\value{
Tibble with the full data for the ACS table.
}
\description{
Enhance a single ACS table and return the full table for geographies of interest
}
\examples{
df1 <- get_acstab("B01001")
df2 <- enhance(df1)
}
