% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_tidy.R
\name{get_enhanced}
\alias{get_enhanced}
\title{Get and enhance a single ACS table and return the full table for geographies of interest}
\usage{
get_enhanced(tabname)
}
\arguments{
\item{tabname}{Census Bureau name for the table (e.g., B01001) as a string.}
}
\value{
Tibble with the full data for the ACS table.
}
\description{
Get and enhance a single ACS table and return the full table for geographies of interest
}
\examples{
df2 <- get_enhanced("B01003")
}
